// src/shared/components/text/text.types.ts
export enum textTypes {
  TITLE_BOLD = "TITLE_BOLD",
  TITLE_REGULAR = "TITLE_REGULAR",
  TITLE_LIGHT = "TITLE_LIGHT",
  SUBTITLE_BOLD = "SUBTITLE_BOLD",
  SUBTITLE_REGULAR = "SUBTITLE_REGULAR",
  SUBTITLE_LIGHT = "SUBTITLE_LIGHT",
  BUTTON_BOLD = "BUTTON_BOLD",
  BUTTON_REGULAR = "BUTTON_REGULAR",
  BUTTON_LIGHT = "BUTTON_LIGHT",
  PARAGRAPH_BOLD = "PARAGRAPH_BOLD",
  PARAGRAPH_REGULAR = "PARAGRAPH_REGULAR",
  PARAGRAPH_LIGHT = "PARAGRAPH_LIGHT",
  PARAGRAPH_SMALL_BOLD = "PARAGRAPH_SMALL_BOLD",
  PARAGRAPH_SMALL_REGULAR = "PARAGRAPH_SMALL_REGULAR",
  PARAGRAPH_SMALL_LIGHT = "PARAGRAPH_SMALL_LIGHT",
}